Rename packages from methods to strategies
