+++ src/main/java/ar/edu/itba/genetic_algorithms/algorithms/crossover_strategies/OnePointCrossover.java
@@ -1,11 +1,11 @@
-package ar.edu.itba.genetic_algorithms.algorithms.crossover_methods;
+package ar.edu.itba.genetic_algorithms.algorithms.crossover_strategies;
 
 import ar.edu.itba.genetic_algorithms.algorithms.api.Chromosome;
 import ar.edu.itba.genetic_algorithms.algorithms.engine.ChromosomePair;
 
 import java.util.Random;
 
-public class UniformCrossover implements CrossoverStrategy {
+public class OnePointCrossover implements CrossoverStrategy {
 
     @Override
     public ChromosomePair crossover(ChromosomePair parents) {
@@ -13,21 +13,17 @@ public class UniformCrossover implements CrossoverStrategy {
         Object[] parentGenes1 = parents.getFirst().getGenes();
         Object[] parentGenes2 = parents.getSecond().getGenes();
         int chromosomeLength = parentGenes1.length;
+        int locus = r.nextInt(chromosomeLength);
         Object[] offspringGenes1 = new Object[chromosomeLength];
         Object[] offspringGenes2 = new Object[chromosomeLength];
 
         for(int i = 0; i < chromosomeLength; i++){
-            if(r.nextBoolean()){
+            if(i < locus){
                 offspringGenes1[i] = parentGenes1[i];
                 offspringGenes2[i] = parentGenes2[i];
             } else {
-                if(r.nextBoolean()){
-                    offspringGenes1[i] = parentGenes1[i];
-                    offspringGenes2[i] = parentGenes2[i];
-                } else {
-                    offspringGenes1[i] = parentGenes2[i];
-                    offspringGenes2[i] = parentGenes1[i];
-                }
+                offspringGenes1[i] = parentGenes2[i];
+                offspringGenes2[i] = parentGenes1[i];
             }
         }
